e2e: Mv exec GA ping readiness probe
Currently, the exec readiness probe tests are owned by SIG-network.
The exec probes is a feature owned by SIG-Compute.
An exec probe does not exercise network-related code paths.
